Reduce a sharp edge in the public API. Before we would keep a reference
internally. Users would be suprised when values on the passed
*rtp.Packet would change.

TODO write Benchmark