@Sean-Der 2020-11-29
@at-wat @OrlandoCo @bshi (anyone else who is good with performance stuff?) What do you think of this change, here are my results so far.

The goal here is to reduce sharp edges and some motivation from https://medium.com/@vCabbage/go-are-pointers-a-performance-optimization-a95840d3ef85

```
# Copy Based
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          75387             15420 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76329             14994 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          72397             15292 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          75758             15383 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76332             16396 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          79160             15475 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          62388             19728 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          77582             15672 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76843             17727 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          75664             14963 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76006             15033 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          78680             15036 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          61647             16346 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76102             15309 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76863             15439 ns/op



# Pointer Based
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          53791             19168 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76670             15054 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76576             14956 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76573             15313 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          79381             15019 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          76624             15312 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          75909             14949 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          79003             16044 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          78114             14904 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          77294             15068 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          81478             16868 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          78312             15229 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          78694             14963 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          75019             15080 ns/op
Benchmark_TrackLocalStaticRTP_SendOnly_WriteRTP-4          71244             15425 ns/op
```