@Sean-Der 2020-11-29
@OrlandoCo one issue is that `pion/srtp` expects a *rtp.Packet, so we probably should address there also!

I am tempted to say `/v4` I just really want to get `Per Transceiver Configuration` done and this is starting to seem like a larger task