@OrlandoCo 2020-11-30
> @OrlandoCo one issue is that `pion/srtp` expects a *rtp.Packet, so we probably should address there also!
> 
> I am tempted to say `/v4` I just really want to get `Per Transceiver Configuration` done and this is starting to seem like a larger task

The main concern was that now by default the `TrackLocalStaticRTP` modify the header, not sure if for some implementations it may cause weird bugs if the depend of the original values of the packets. We can just for now just change the behaviour in the writer to make a copy there, and later make the change. But reduce allocations in hot paths like writing packets would be great!