@bshi 2020-11-30
Using https://godoc.org/golang.org/x/perf/cmd/benchstat

```
$ benchstat /tmp/copy.txt /tmp/ptr.txt
name                                      old time/op  new time/op  delta
_TrackLocalStaticRTP_SendOnly_WriteRTP-4  15.6µs ±14%  15.2µs ± 6%  -2.74%  (p=0.044 n=14+13)
```

On the surface seems like the safety improvement comes with a slight improvement in timing benchmarks (just barely statistically significant P<0.05).  It would be useful to see how this impacts allocations via `b.ReportAllocs()` in your benchmark function or `-benchmem` flag.