@Sean-Der 2020-11-30
Ah I completely missed your point originally @OrlandoCo I simplified this a lot.

I opened https://github.com/pion/webrtc/issues/1560 and we should do a deep dive into performance/API of Pion across the board. We just haven't been at a point of having enough features where I can focus on this. Anymore it seems like performance has been the thing most users have reached out about (after Interceptors)